Suppress compiler warning
In trying to build this on Arch Linux, I had to add this option to suppress compiler warnings as they were being treated as errors. However, this is at most a temporary fix and a move towards Modern CMake design patterns should be considered to streamline the build process and better implement the library.